Transfer MailSenderService to MailSender.lib
